addded view and model files
