+++ src/index.html
@@ -29,49 +29,17 @@
 
         <div class="city-weather wraper__city-weather wraper" data-action="city-weather">
 
-            <!-- <div class="city-list wraper__city-list">
-
-                <div class="favorit">
-                    <button class="btn-favorite"></button>
-                </div>
-
-                <h1 class="temp">22°c</h1>
-                <h2 class="cityName">Moscow</h2>
-
-                <div class="hum-wind city-list__hum-wind">
-                    <div class="humidity-wrapper hum-wind__humidity-wrapper">
-                        <img class="humidity-icon icon" src="img/humidity.png" alt="icon humidity">
-                        <div class="text">
-                            <p class="humidity">50%</p>
-                            <p class="hum-wind-text">Humidity</p>
-                        </div>
-                    </div>
-
-                    <div class="wind-wrapper hum-wind__wind-wrapper">
-                        <img class="wind-icon icon" src="img/wind.png" alt="icon wind">
-                        <div class="text">
-                        <p class="wind">15 km/h</p>
-                        <p class="hum-wind-text">Wind</p>
-                    </div>
-                </div>
-            </div>
-
-            <div class="button">
-                <button type="button" data-action="delete" class="button-delete btn">delete</button>
-            </div>
-        </div> -->
-
         </div>
     </div>
 
 
     <div id="popup" class="popup" data-action="popup">
-        <!-- <div class="popup__content">
-            <div class="popup__close">&#10006</div>
-        </div> -->
+
     </div>
 
-    <script src="index.js"></script>
+    <script type="module" src="js/view.js"></script>
+    <script type="module" src="js/model.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 
 </html>
