+++ src/style.css
@@ -335,3 +335,7 @@ input::placeholder {
 .loader-visible {
     display: block;
 }
+
+.loader-none {
+    display: none;
+}
